Initial stubs for a push/pull world
