+++ src/EditorFeatures/Core/Diagnostics/AbstractAggregateDiagnosticsTaggerProvider.SingleDiagnosticKindTaggerProvider.cs
@@ -23,151 +23,11 @@ using Microsoft.CodeAnalysis.Text.Shared.Extensions;
 using Microsoft.CodeAnalysis.Workspaces;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Text.Tagging;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Diagnostics;
 
-internal partial class AbstractAggregateDiagnosticsTaggerProvider<TTag>
+internal abstract partial class AbstractPushOrPullDiagnosticsTaggerProvider<TTag>
 {
-    /// <summary>
-    /// Low level tagger responsible for producing specific diagnostics tags for some feature for some particular <see
-    /// cref="DiagnosticKind"/>.  It is itself never exported directly, but it it is used by the <see
-    /// cref="AbstractAggregateDiagnosticsTaggerProvider{TTag}"/> which aggregates its results and the results for all the
-    /// other <see cref="DiagnosticKind"/> to produce all the diagnostics for that feature.
-    /// </summary>
-    private sealed class SingleDiagnosticKindTaggerProvider : AsynchronousTaggerProvider<TTag>
-    {
-        private readonly DiagnosticKind _diagnosticKind;
-        private readonly IDiagnosticService _diagnosticService;
-        private readonly IDiagnosticAnalyzerService _analyzerService;
-
-        private readonly AbstractAggregateDiagnosticsTaggerProvider<TTag> _callback;
-
-        protected override ImmutableArray<IOption> Options => _callback.Options;
-
-        public SingleDiagnosticKindTaggerProvider(
-            AbstractAggregateDiagnosticsTaggerProvider<TTag> callback,
-            DiagnosticKind diagnosticKind,
-            IThreadingContext threadingContext,
-            IDiagnosticService diagnosticService,
-            IDiagnosticAnalyzerService analyzerService,
-            IGlobalOptionService globalOptions,
-            ITextBufferVisibilityTracker? visibilityTracker,
-            IAsynchronousOperationListener listener)
-            : base(threadingContext, globalOptions, visibilityTracker, listener)
-        {
-            _callback = callback;
-            _diagnosticKind = diagnosticKind;
-            _diagnosticService = diagnosticService;
-            _analyzerService = analyzerService;
-        }
-
-        protected sealed override TaggerDelay EventChangeDelay => TaggerDelay.Short;
-        protected sealed override TaggerDelay AddedTagNotificationDelay => TaggerDelay.OnIdle;
-
-        /// <summary>
-        /// When we hear about a new event cancel the costly work we're doing and compute against the latest snapshot.
-        /// </summary>
-        protected sealed override bool CancelOnNewWork => true;
-
-        protected sealed override bool TagEquals(TTag tag1, TTag tag2)
-            => _callback.TagEquals(tag1, tag2);
-
-        protected sealed override ITaggerEventSource CreateEventSource(ITextView? textView, ITextBuffer subjectBuffer)
-        {
-            // OnTextChanged is added for diagnostics in source generated files: it's possible that the analyzer driver
-            // executed on content which was produced by a source generator but is not yet reflected in an open text
-            // buffer for that generated file. In this case, we need to update the tags after the buffer updates (which
-            // triggers a text changed event) to ensure diagnostics are positioned correctly.
-            return TaggerEventSources.Compose(
-                TaggerEventSources.OnDocumentActiveContextChanged(subjectBuffer),
-                TaggerEventSources.OnWorkspaceRegistrationChanged(subjectBuffer),
-                TaggerEventSources.OnDiagnosticsChanged(subjectBuffer, _diagnosticService),
-                TaggerEventSources.OnTextChanged(subjectBuffer));
-        }
-
-        protected sealed override Task ProduceTagsAsync(
-            TaggerContext<TTag> context, DocumentSnapshotSpan spanToTag, int? caretPosition, CancellationToken cancellationToken)
-        {
-            return ProduceTagsAsync(context, spanToTag, cancellationToken);
-        }
-
-        private async Task ProduceTagsAsync(
-            TaggerContext<TTag> context, DocumentSnapshotSpan documentSpanToTag, CancellationToken cancellationToken)
-        {
-            if (!_callback.IsEnabled)
-                return;
-
-            var diagnosticMode = GlobalOptions.GetDiagnosticMode(InternalDiagnosticsOptions.NormalDiagnosticMode);
-            if (!_callback.SupportsDiagnosticMode(diagnosticMode))
-                return;
-
-            var document = documentSpanToTag.Document;
-            if (document == null)
-                return;
-
-            var snapshot = documentSpanToTag.SnapshotSpan.Snapshot;
-
-            var workspace = document.Project.Solution.Workspace;
-
-            // See if we've marked any spans as those we want to suppress diagnostics for.
-            // This can happen for buffers used in the preview workspace where some feature
-            // is generating code that it doesn't want errors shown for.
-            var buffer = snapshot.TextBuffer;
-            var suppressedDiagnosticsSpans = (NormalizedSnapshotSpanCollection?)null;
-            buffer?.Properties.TryGetProperty(PredefinedPreviewTaggerKeys.SuppressDiagnosticsSpansKey, out suppressedDiagnosticsSpans);
-
-            var sourceText = snapshot.AsText();
-
-            try
-            {
-                var diagnostics = await _analyzerService.GetDiagnosticsForSpanAsync(
-                    document,
-                    documentSpanToTag.SnapshotSpan.Span.ToTextSpan(),
-                    diagnosticKind: _diagnosticKind,
-                    cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                var requestedSpan = documentSpanToTag.SnapshotSpan;
-
-                foreach (var diagnosticData in diagnostics)
-                {
-                    if (_callback.IncludeDiagnostic(diagnosticData))
-                    {
-                        // We're going to be retrieving the diagnostics against the last time the engine
-                        // computed them against this document *id*.  That might have been a different
-                        // version of the document vs what we're looking at now.  But that's ok:
-                        // 
-                        // 1) GetExistingOrCalculatedTextSpan will ensure that the diagnostics spans are
-                        //    contained within 'editorSnapshot'.
-                        // 2) We'll eventually hear about an update to the diagnostics for this document
-                        //    for whatever edits happened between the last time and this current snapshot.
-                        //    So we'll eventually reach a point where the diagnostics exactly match the
-                        //    editorSnapshot.
-
-                        var diagnosticSpans = _callback.GetLocationsToTag(diagnosticData)
-                            .Select(loc => loc.UnmappedFileSpan.GetClampedTextSpan(sourceText).ToSnapshotSpan(snapshot));
-                        foreach (var diagnosticSpan in diagnosticSpans)
-                        {
-                            if (diagnosticSpan.IntersectsWith(requestedSpan) && !IsSuppressed(suppressedDiagnosticsSpans, diagnosticSpan))
-                            {
-                                var tagSpan = _callback.CreateTagSpan(workspace, diagnosticSpan, diagnosticData);
-                                if (tagSpan != null)
-                                    context.AddTag(tagSpan);
-                            }
-                        }
-                    }
-                }
-            }
-            catch (ArgumentOutOfRangeException ex) when (FatalError.ReportAndCatch(ex))
-            {
-                // https://devdiv.visualstudio.com/DefaultCollection/DevDiv/_workitems?id=428328&_a=edit&triage=false
-                // explicitly report NFW to find out what is causing us for out of range. stop crashing on such
-                // occasions
-                return;
-            }
-        }
-
-        private static bool IsSuppressed(NormalizedSnapshotSpanCollection? suppressedSpans, SnapshotSpan span)
-            => suppressedSpans != null && suppressedSpans.IntersectsWith(span);
-    }
 }
